+++ tests/testthat/test-gsnPathways.R
@@ -14,6 +14,10 @@ test_that( "gsnPathways works", {
   expect_equal( gsnPathways( .STLF.GSN.cp, "TEST" ), expected = c("A", "B", "C", "D", "E" ) )
   expect_error( gsnPathways( .STLF.GSN.cp, "TEST2" ) <- c("A", "B", "C", "D", "E", "F" ) )
 
+  # Make sure factors work. (They didn't in version 0.1.6.5)
+  expect_no_error( gsnPathways( .STLF.GSN.cp, "TEST_FACTOR" ) <- factor( c("A", "B", "C", "D", "E" ) ) )
+
+  # Assign a data frame to multiple columns.
   .assign <- data.frame( TEST3 = c("A", "B", "C", "D", "E" ), TEST4 = c(1,2,3,4,5) )
   expect_no_error( gsnPathways( .STLF.GSN.cp, "TEST3", "TEST4" ) <- .assign )